destroyアクションへのアクセス権を設定

+++ app/controllers/items_controller.rb
@@ -37,6 +37,7 @@ class ItemsController < ApplicationController
   end
 
   def destroy
+    redirect_to root_path unless current_user == @item.user
     redirect_to root_path if @item.destroy
   end
